+++ little_brother/constants.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (C) 2019  Marcus Rickert
+# Copyright (C) 2019-2021  Marcus Rickert
 #
 # See https://github.com/marcus67/little_brother
 # This program is free software; you can redistribute it and/or modify
@@ -19,6 +19,7 @@ import os.path
 
 DIR_NAME = 'little-brother'
 APPLICATION_USER = 'little-brother'
+APPLICATION_NAME = 'LittleBrother'
 
 DEFAULT_RULE_SET_PRIORITY = 1